py3 fix

+++ qrcodenative.py
@@ -929,7 +929,7 @@ class QRRSBlock:
       if rsBlock == None:
          raise Exception("bad rs block @ typeNumber:" + typeNumber + "/errorCorrectLevel:" + errorCorrectLevel)
 
-      length = len(rsBlock) / 3
+      length = int(len(rsBlock) / 3)
 
       list = []
